totlog decimal allowed

+++ Models/TimeOnTools/TOTLog.cs
@@ -15,7 +15,7 @@ namespace Models.TimeOnTools
         public string Twr { get; set; }
 
         public string? WorkScope { get; set; }
-        public long? ManHours { get; set; }
+        public double? ManHours { get; set; }
         public DateTime? StartOfWork { get; set; }
         public TimeSpan? TimeRequested { get; set; }
         public TimeSpan? TimeSigned { get; set; }
